requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
PyQt5